scripts: create swap file only once

+++ scripts/linux-gnu/swap.sh
@@ -4,8 +4,10 @@ set -xe
 
 swapfile=/swap/file
 
-dd if=/dev/zero of=$swapfile bs=1GB count=14
-chmod 0600 $swapfile
+if [ ! -f $swapfile ]; then
+    dd if=/dev/zero of=$swapfile bs=1GB count=14
+    chmod 0600 $swapfile
+fi
 
 mkswap $swapfile
 swapon $swapfile
